+++ install-src.sh
@@ -34,22 +34,6 @@ install -d "$targetdir/src/p"
 install -v \
     src/p/flexdll-cross-0_99-a01-arm64.patch \
     src/p/ocaml-common-4_12-a01-alignfiletime.patch \
-    src/p/ocaml-common-4_14-a01-alignfiletime.patch \
-    src/p/ocaml-common-4_14-a02-nattop.patch \
-    src/p/ocaml-common-4_14-a03-keepasm.patch \
-    src/p/ocaml-common-4_14-a04-xdg.patch \
-    src/p/ocaml-common-4_14-a05-msvccflags.patch \
-    src/p/ocaml-common-4_14-a06-linearclosures.patch \
-    src/p/ocaml-common-4_14-a07-arm32dedup.patch \
-    src/p/ocaml-common-4_14-a08-declspecdllimport.patch \
-    src/p/ocaml-common-4_14-a09-debuggasx86.patch \
-    src/p/ocaml-common-4_14-a10-bx32.patch \
-    src/p/ocaml-common-4_14-a11-relocatable-version-build.patch \
-    src/p/ocaml-common-4_14-a12-relocatable-launcher.patch \
-    src/p/ocaml-common-4_14-a13-relocatable-stdlib.patch \
-    src/p/ocaml-common-4_14-a14-relocatable-runtime-search.patch \
-    src/p/ocaml-common-4_14-a15-relocatable-install.patch \
-    src/p/ocaml-common-4_14-a16-bigarray-msvc18.patch \
     src/p/ocaml-common-4_14_0-a01-fmatest.patch \
     src/p/ocamldebug_gcc_clang-common-4_14-a01-compileflags.patch \
     src/p/ocaml-cross-4_12-a01.patch \
@@ -61,6 +45,8 @@ install -v \
     src/p/ocaml-cross-4_14_2-a02-arm32.patch \
     src/p/ocaml-cross-5_00_a02-arm32.patch \
     "$targetdir/src/p"
+
+install -v src/p/ocaml-common-4_14-* "$targetdir/src/p"
 
 install -d "$targetdir/env"
 install -v \